refactor: httpMessage 패키지 추가
