+++ tomcat/src/main/java/org/apache/coyote/http11/urihandler/LoginUriHandler.java
@@ -6,7 +6,7 @@ import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.UriResponse;
-import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.httpmessage.request.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
