+++ tomcat/src/main/java/org/apache/coyote/http11/urihandler/RootUriHandler.java
@@ -3,7 +3,7 @@ package org.apache.coyote.http11.urihandler;
 import java.util.regex.Pattern;
 import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.UriResponse;
-import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.httpmessage.request.HttpRequest;
 
 public class RootUriHandler implements UriHandler {
