+++ tomcat/src/test/java/org/apache/coyote/http11/httpmessage/request/HttpRequestTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.httpmessage.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
